requirements.txt
# AI Prediction API — Python Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
aiomysql==0.2.0
pymysql==1.1.1

# ML / Data
tensorflow==2.17.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.7.2

# HTTP (for Groq/Ollama calls)
httpx==0.27.2

# Environment
python-dotenv==1.0.1

